requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
Deprecated==1.2.14
et-xmlfile==1.1.0
frozenlist==1.4.1
g4f==0.3.2.0
greenlet==3.0.3
idna==3.7
mss==9.0.1
multidict==6.0.5
nest-asyncio==1.6.0
nodriver==0.29rc2
openpyxl==3.1.2
pyaes==1.6.1
pyasn1==0.5.1
pycryptodome==3.20.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.27
Telethon==1.34.0
typing_extensions==4.10.0
urllib3==2.2.1
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
